Move common UI elements from :app to :ui:common
