+++ app/src/main/java/de/danoeh/antennapod/activity/MainActivity.java
@@ -80,8 +80,8 @@ import de.danoeh.antennapod.ui.screen.queue.QueueFragment;
 import de.danoeh.antennapod.ui.screen.rating.RatingDialogManager;
 import de.danoeh.antennapod.ui.screen.subscriptions.SubscriptionFragment;
 import de.danoeh.antennapod.ui.statistics.StatisticsFragment;
-import de.danoeh.antennapod.ui.view.BottomSheetBackPressedCallback;
-import de.danoeh.antennapod.ui.view.LockableBottomSheetBehavior;
+import de.danoeh.antennapod.ui.common.BottomSheetBackPressedCallback;
+import de.danoeh.antennapod.ui.common.LockableBottomSheetBehavior;
 import org.apache.commons.lang3.ArrayUtils;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
