+++ app/src/main/res/layout-sw720dp/main.xml
@@ -47,7 +47,7 @@
                 android:background="?android:attr/colorBackground"
                 android:elevation="8dp"
                 android:visibility="gone"
-                app:layout_behavior="de.danoeh.antennapod.ui.view.LockableBottomSheetBehavior" />
+                app:layout_behavior="de.danoeh.antennapod.ui.common.LockableBottomSheetBehavior" />
 
         </androidx.coordinatorlayout.widget.CoordinatorLayout>
