+++ app/src/main/res/layout/feeditem_fragment.xml
@@ -204,7 +204,7 @@
 
     </LinearLayout>
 
-    <de.danoeh.antennapod.ui.view.NestedScrollableHost
+    <de.danoeh.antennapod.ui.common.NestedScrollableHost
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_below="@id/header"
@@ -216,7 +216,7 @@
             android:layout_height="match_parent"
             android:foreground="?android:windowContentOverlay" />
 
-    </de.danoeh.antennapod.ui.view.NestedScrollableHost>
+    </de.danoeh.antennapod.ui.common.NestedScrollableHost>
 
     <FrameLayout
         android:layout_width="match_parent"
