+++ app/src/main/res/layout/item_description_fragment.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<de.danoeh.antennapod.ui.view.NestedScrollableHost
+<de.danoeh.antennapod.ui.common.NestedScrollableHost
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
@@ -13,4 +13,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-</de.danoeh.antennapod.ui.view.NestedScrollableHost>
+</de.danoeh.antennapod.ui.common.NestedScrollableHost>
