--- app/src/main/res/values/attrs.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <declare-styleable name="NestedScrollableHost">
-        <attr name="scrollDirection" format="enum">
-            <enum name="both" value="0"/>
-            <enum name="vertical" value="1"/>
-            <enum name="horizontal" value="2"/>
-        </attr>
-        <attr name="preferHorizontal" format="integer"/>
-        <attr name="preferVertical" format="integer"/>
-    </declare-styleable>
-</resources>
