+++ ui/common/src/main/java/de/danoeh/antennapod/ui/common/ItemOffsetDecoration.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.ui.view;
+package de.danoeh.antennapod.ui.common;
 
 import android.content.Context;
 import android.graphics.Rect;
