+++ ui/common/src/main/res/layout/empty_view_layout.xml
