+++ ui/common/src/main/res/values/attrs.xml
@@ -11,4 +11,14 @@
     <attr name="icon_green" format="color" />
     <attr name="icon_purple" format="color" />
     <attr name="icon_gray" format="color" />
+
+    <declare-styleable name="NestedScrollableHost">
+        <attr name="scrollDirection" format="enum">
+            <enum name="both" value="0"/>
+            <enum name="vertical" value="1"/>
+            <enum name="horizontal" value="2"/>
+        </attr>
+        <attr name="preferHorizontal" format="integer"/>
+        <attr name="preferVertical" format="integer"/>
+    </declare-styleable>
 </resources>
